requirements.txt
Flask
gTTS
Werkzeug
itsdangerous
Jinja2
click
supabase
python-dotenv